Extended the clock to be able to hide second pointer and move minute pointer smoother when that’s the case. Also made it notify on each minute change. (#2)
Minute changes are now acted upon and sun position and intensity is adjusted on these. Sun calculations was choppy earlier, so also cached a full day of values and smoothened them out. So when a level (clockwise) is sped up, the sun will move quite nicely (limited levels to 24hrs clock time)